+++ src/App.jsx
@@ -1,20 +1,20 @@
-import { useState } from 'react'
+import { } from 'react'
 import HeaderComponent from './HeaderComponent.jsx'
 import Search from './Search.jsx'
 import Greenband from './Greenband.jsx'
-import Card from './Card.jsx'
+import CardHolder from './CardHolder.jsx'
 
 import './App.css'
 
 function App() {
-  
+
 
   return (
     <>
-   <HeaderComponent/>
-    <Greenband/>
-    <Search/>
-    <Card/>
+      <HeaderComponent />
+      <Greenband />
+      <Search />
+      <CardHolder />
     </>
   )
 }
