--- src/Card.jsx
@@ -1,96 +1,0 @@
-import React from 'react'
-
-function Card() {
-  return (
-    <>
-      <section className='Cart-down'>
-      <span className='cardd'>
-         <h3 >
-            Complete Blood Count (CBC)
-         </h3>
-         <p>28 Tests included</p><hr/>
-         <p>Mrp $399</p>
-         <button>Add To Cart</button>
-      </span>
-      <span className='cardd-1'>
-      <h3 >
-            Thyroid Profile (Total T3,Total T4,TSH)
-         </h3>
-         <p>4 Tests included</p><hr/>
-         <p>Mrp $439</p>
-         <button>Add To Cart</button>
-      </span>
-      <span className='cardd-2'>
-      <h3 >
-            Lipid <br/>Profile 
-         </h3>
-         <p>28 Tests included</p><hr/>
-         <p>Mrp $399</p>
-         <button>Add To Cart</button>
-      </span>
-      <span className='cardd-3'>
-      <h3 >
-            Complete Blood Count (CBC)
-         </h3>
-         <p>28 Tests included</p><hr/>
-         <p>Mrp $399</p>
-         <button>Add To Cart</button>
-      </span>
-      <span className='cardd-4'>
-      <h3 >
-            Complete Blood Count (CBC)
-         </h3>
-         <p>28 Tests included</p><hr/>
-         <p>Mrp $399</p>
-         <button>Add To Cart</button>
-      </span>
-      </section>
-
-      <section className='Cart-down'>
-      <span className='cardd'>
-         <h3 >
-            Complete Blood Count (CBC)
-         </h3>
-         <p>28 Tests included</p><hr/>
-         <p>Mrp $399</p>
-         <button>Add To Cart</button>
-      </span>
-      <span className='cardd-1'>
-      <h3 >
-            Complete Blood Count (CBC)
-         </h3>
-         <p>28 Tests included</p><hr/>
-         <p>Mrp $399</p>
-         <button>Add To Cart</button>
-      </span>
-      <span className='cardd-2'>
-      <h3 >
-            Complete Blood Count (CBC)
-         </h3>
-         <p>28 Tests included</p><hr/>
-         <p>Mrp $399</p>
-         <button>Add To Cart</button>
-      </span>
-      <span className='cardd-3'>
-      <h3 >
-            Complete Blood Count (CBC)
-         </h3>
-         <p>28 Tests included</p><hr/>
-         <p>Mrp $399</p>
-         <button>Add To Cart</button>
-      </span>
-      <span className='cardd-4'>
-      <h3 >
-            Complete Blood Count (CBC)
-         </h3>
-         <p>28 Tests included</p><hr/>
-         <p>Mrp $399</p>
-         <button>Add To Cart</button>
-      </span>
-      </section>
-    </>
-  
-  )
-}
-
-export default Card
